[IO] Rewrite of IO to use C's Standard IO
In order to reduce the complexity of the implementation and make
testing easier, this PR contains a complete rewrite of the IOWrapper
and the StandardIO class, to use C's Standard IO (`stdio.h`)

Notably, the class is not complete since some operations (e.g. renaming,
deleting, etc.) are not required at this time. The StandardIO class is
covered by tests